chore(probes): rename probe_group and adjust code
